requirements.txt
langchain
openai
faiss-cpu
gradio
python-dotenv
tiktoken
PyPDF2
